+++ src/app/packages/core/models/index.ts
@@ -4,3 +4,4 @@ export { MenuItem } from './menu-item.model';
 export { IGeneralConfig } from './general.config';
 export { Case, CaseState, CaseTimeState, CaseFacetsState, CaseMapsState, CaseMapState, defaultMapType } from './case.model';
 export { MapsLayout } from './maps-layout';
+export { Context } from './context.model';
