+++ src/app/packages/menu-items/cases/models/context.model.ts
@@ -1,16 +1,1 @@
-export type Context = {
-	id: string,
-	name: string,
-	time: {
-		type:string,
-		from: Date,
-		to: Date
-	},
-	facets: {
-		SensorName: string,
-		SensorType: 'SAR' | 'VIS' | 'IR' | 'Satellite',
-		Stereo: boolean,
-		Resolution: number
-	},
-	region: {}
-}
+export { Context } from '@ansyn/core';
